requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
requests
python-decouple
psycopg2-binary
gunicorn
